Move plot-functions to plotting module
